requirements.txt
colorama==0.4.6
exceptiongroup==1.2.2
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.2.0
pyserial==3.5
pytest==7.4.4
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
